Copy changes to 'dx_vif.go' and 'structure.go' from the 'issue-3248' branch.
